Move Operations tests to separate file
